Implement SO_REUSEPORT sockets
We've found that getting raw FDs, setting REUSEPORT on them, and
wrapping them in a FilePacketConn has significant implications for
packet drop rate. If multiple goroutines each read from their own
REUSEPORT socket, we can cut drop rate down to zero on boxes where it
would otherwise hover around 2-3%. This corroborates findings that
similar projects have had in C, but I wasn't sure that it would work in
Go until we tried it.

At the moment the implementation only supports Linux. Theoretically it
could support BSDs as well but we just don't need it. If x/sys/unix adds
the SO_REUSEPORT constant on Linux it would be easier.