Move metric flush invocations to s.Flush
The flush function will need to have access to raw metric structures in
the near future, so aggregating the metrics array has to happen here,
not in the worker. This generates more garbage since we're repeatedly
appending, but since we only do this on flush, I'm not hugely concerned
about it.