Delay attaching custom tags until flush time
We want the metrics we store to only have the tags that were specified
by the client. That way, sampler structures do not get polluted with our
own tags.

This will become important when we start forwarding the sampler
structures to upstream veneur instances. If the structure we forward has
our own tags on it, then the global instance might attach its own tags
(which might be the same ones again), and it all just gets very icky.