Prevent veneurs from forwarding their own metrics
Veneur attaches its host-specific tags to the metrics that it internally
generates. If these get forwarded to another veneur, the tags might be
duplicated (when that veneur adds its own host-local tags at flush
time). We could try to do some clever deduplication, but that's annoying
and expensive. We can afford to sacrifice globalization for veneur
itself to avoid this data inconsistency.