Copy tag slices in sampler flushing
We actually have an aliasing bug where all the DDMetrics returned from a
histogram are using the same slice of tags, and then the server-wide
tags are being appended to that slice over and over again. Right now
it's harmless, but we should probably ensure that no future bugs slip in
this way.