Perform t-digest merges in place
Before this commit, each t-digest contained a temporary buffer to
perform merges into, which approximately doubles their memory
consumption. This commit introduces an in-place merge that makes this
extra buffer unnecessary. It's a slight performance regression (adding
samples to the histogram, which amortizes the cost of merges across all
samples, takes about 50% longer on my machine) but nearly halves memory
consumption of this object.