add counters that can be globally accumulated
Counters are emitted locally by each host, which can increase tag cardinality
substantially for a metric emitted by many hosts. This change allows counters to
be flushed to a global veneur instance, which aggregates and flushes the total
to DataDog.

This change assumes that all veneur instances use the same flush interval.