config: omit_empty_hostname: Allow not sending hostname to Datadog
By default, Veneur drops the host tag for global metrics, and always
tags local metrics. In Bluecore's case, we are currently using a
single Veneur instance to aggegrate metrics from an existing system,
so our metrics are "local", but the host tag doesn't make sense.

We could specify a generic host tag, like "veneur", but we think
it is less confusing to omit it. Additionally at some point we might
need to use a "tree" of Veneur instances if we are lucky enough to
need to scale up to large numbers of metrics.