Support statsd over TCP, encrypted or authenticated TLS connections
This adds configuration options to enable listening on a TCP port in
one of three modes:

* Plain TCP: unencrypted and unauthenticated
* TLS: encrypted but allows any client
* TLS with client certificates: encrypted and authenticated

By default, Veneur (and statsd, and dogstatsd) accepts stats from any
client that can send UDP packets to the correct port. In some cases
it is useful to only accept data from authenticated clients. An
"easy" way to do this securely is to use an encrypted TLS connection,
and require the client to present a certificate signed by a trusted
authority. This reuses the existing authentication mechanism in TLS,
which is hopefully secure.

server_test: Add fixture.flush to reduce test duplication