sentry.go: sentryHook.Fire: Don't hang for Fatal logs without Sentry
If s.c is nil, s.c.Capture returns a nil chan. Reading from it blocks
forever. This means if someone logs at logrus Fatal or Panic levels
without Sentry configured, the caller will hang instead of causing
the process. This is the same bug as commit d04711e9c2, just in a
different place.

It turns out the unit tests were depending on this behaviour, at
least on Mac OS X. Fixing this requires fixing Server.Shutdown to
stop listening on the socket, to avoid address in use errors.