server_test.go: Use new ports to avoid address in use errors (#134)
Server.Shutdown does not stop the metrics or trace listeners. This
causes address in use errors that cause a call to logrus.Fatal.
Currently this is being masked due to a bug in sentryHook.Fire which
causes it to hang instead of exit. However, fixing that bug requires
this to be fixed.

server.go: Document that the way we shut down the TCP listener is not
completely correct.