server.ReadTCPSocket: Remove useless defer; fix memory leak
ReadTCPSocket is called in a goroutine that is created in
Server.Start, and ConsumePanic is already called there. This extra
call does nothing. Even worse, it is in a loop, so it allocates a new
closure each time, leaking memory.

This was added by me in commit 7af6d66ede. I'm pretty sure this was a
refactoring error I introduced and overlooked.