Server.handleTCPGoroutine: Close idle connections after 10 mins
When established TCP connections break unexpectedly, readers can hang
indefinitely. This causes a long running server to leak memory and
goroutines, as more and more dead idle sockets build up. Time out and
close connections after a default of 10 minutes.

Always call tlsConn.Handshake: The way connections are created it is
always required, and it will also work even if it is not. This
simplifies the code a bit.