Add Bazel build configs for Veneur.
This allows Veneur to be easily used as an external versioned dependency
instead of having to vendor it, or fetch an unversioned binary from S3.

In the interest of maintainability, we may want to migrate `Dockerfile`
to use Bazel so `vendor/vendor.json` can be deleted. Most of the
existing Docker commands could be replaced with Bazel external dependencies.

Tested by running `bazel build //...` and `bazel test //...`; both succeed.