flusher_test.go: TestFlushTracesDatadog: Fix occasional test failures
* Parse the expected message once, rather than on each trace POST
* Compare the expected POST body once, in the main test.
* Remove duplicate assert.NoError(...)

On my machine this test fails about once every 3 runs. It seems the
reason is that the test HTTP server endpoint gets called more than
once: the first time by the actual test, and the second time by the
server's normal timed flush getting called. The second call fails,
since the expected message has already been parsed, so it gets an
EOF when reading from jsn, and fails assert.Equal since the post does
not contain the expected traces. With this change, the test passed
after running it 100 times in a loop.