Refactor sending of spans with a client structure
This introduces a Client structure, which can back onto buffered and
unbuffered streaming connections, as well as regular UDP datagram
connections.

That client ensures that user code will never be blocked when sending
a span, unless the code explicitly requests it. Clients serialize the
packets onto a backend, which takes care of reconnecting in case a
connection gets interrupted.