flock a lockfile next to UNIX domain sockets
From https://gavv.github.io/blog/unix-socket-reuse/, this should
ensure that we know at all times if a veneur can re-use an existing
UNIX domain socket file: If the lock file can be acquired, we can do
whatever we want with the existing socket (i.e., delete it). If not,
that means another veneur is running and we can't do anything with it.