Make LS Span ingestion feed the constructed span into the LS tracer
I'd misunderstood what the Ingest function should do: It's meant to
directly inject the constructed span into the LS tracer. This commit
removes a useless cast, which was wrong also: The constructed span can
never traverse the client (so it can never be cast into an SSF span),
and if the cast fails (as it should), the value that we were operating
on was the zero value, instead of the original span.

So, this change restores the previous (correct) behavior.