Correctly split Server.TraceClient initialization
We have to set the variable in NewFromConfig, otherwise the main
function has no way of telling that we're supposed to use the internal
client globally.

This change splits that client initialization up into two steps:
Creation of the (unstarted) trace client in NewFromConfig, starting of
the client in Start.