Histo value: serialize with gob instead of JSON
After reseraching a little more, it seems like gob is the best option
for serializing the Histo value over the wire.  The data is specifically
just being passed to another Go program, and its effectively just an RPC
within the project.  Not to mention the tdigest type already has Gob
encoding built-in, so there are really minimal changes.