Eliminate client test race conditions: retry on ErrWouldBlock
This commit introduces new client test functions, mustRecord and
mustFlush, that ensure a client (if expected) has actually recorded or
flushed a metric. Since the client pump errs towards dropping a span
if there's a hint of a blockage, this triggers on average once every
few test runs but it's really annoying if it does.