Document the `ssf_destination_address` option
The changes to `veneur-proxy` in #338 make the `ssf_destination_address`
a required configuration parameter for the proxy.  This is documented in
"example_proxy.yaml", but there was no information about it in the
`veneur-proxy` README.

This just adds a short description about the configuration option, and
updates the description for `stats_address` to be what I understand it
to now be used for (after #338).