Simplify stream error handling
Previous code wasn't treating io.EOF as a special value in the way that
gRPC expects it to be. This simplifies the error handling a bit, as well
as alllowing us to continue to log potentially interesting non-EOF
errors after the first error has been seen on a given stream instance.