Stronger sequence guarantees in gRPC idle test
It was possible for this test to re-enter the READY state, but have the
test goroutine progress to the point of calling Ingest() before the
state machine manager goroutine got rescheduled and created the stream
for use, resulting in a nil pointer panic.

By adding a reconnectWithin() call and forcing the stream state marker
to bad initially, it should now be impossible for the test goroutine to
proceed before the stream object gets created.