Add forwardrpc (forward gRPC service definition)
The forwardrpc package defines the common gRPC service to send metrics
from one Veneur to another.  It uses the metricpb.(*Metric) type to
actually serialize the metrics, and defines a single RPC to send many of
them in a batch.

I considered implementing this as a client-side streaming RPC, but that
honestly seems like a bit of a premature optimization.  From what I
could tell it would seriously complicate the code for a pretty
unmeasurable performance gain (at least right now).  It may be that this
should be replaced in the future if profiling shows that it would be
fast.

I also added a forwardtest package, which is effectively httptest but
specifically for the Forward service.  This is later used in a couple
different places to greatly simplify test cases around handling of these
RPCs.