Add a gRPC import server (importsrv)
The importsrv package defines a gRPC server that implements the
forwardrpc.Forward service.  It receives batches of metrics, and
sends them on to a set of sinks based on their key.  To facilitate easy
testing of this RPC, I made the server accept an interface
MetricIngester, which is what is used to forward the metrics.  Worker
will later implement this interface, allowing for them to be passed
directly into this server.

I also added a new function to generate a MetricKey based on a
metricpb.(*Metric).  This is used for the hashing in the import server.