Add a proxy gRPC server (proxysrv)
The proxysrv package includes a gRPC server that implements the
forwardrpc.Forward service.  This basically implements the same logic as
the current proxy server, but in a separate package (and obviously with
gRPC).  It uses a consistent hash ring for forwarding metrics, and then
forwards to each one over gRPC.

I added tests as appropriate using the forwardtest mock gRPC server.