Add the gRPC proxy server to the Proxy
The main Proxy now initializes a gRPC server if one exists.  This is
implemented in an extremelly similar way to Server.  It initializes a
proxysrv with the hash ring (specifically for gRPC), and then
initializes the gRPC server.

I added a completely new consul service in the configuration file for
gRPC.  As we'll probably be running the gRPC listener on a separate
port, I thought this made the most sense.