Add support for forwarding over gRPC
This change adds new methods to Server to flush metrics over gRPC if a
configuration option is set.  The implementation is super similar to
the `flushForward` method, except it obviously uses metricpb.(*Metric)'s
as opposed to JSONMetrics.  I also tried to reduce a lot of code
duplication by using a common interface for exporting each different
metric type.

I wrote a couple of associated tests cases to verify that this behavior
works as expected.