Remove usages of the go-multierror package
To make the diff for #407 look a little less insane, I'm going to remove
usages of the go-multierror package.  I replaced it with just using an
array of errors and a custom fuction to merge them, which I figure works
fine as well.  I also reverted all changes to Gopkg.toml and Gopkg.lock
as well as the vendor directory, which should hopefully shrink this diff
by a ton.