Hopefully fix TestReconnectBufferedUNIX
There was one bit of unchecked non-determinism left: When flushing for
the error case, the flush might not get kicked off correctly, which
could cause the subsequent (post-reconnect) flush to send with the
broken connection, timing out the test.

This change makes the kickoff for the "broken" flush retry, which has
so far allowed the test to run hundreds of times without issue.