Pre-initialize a gRPC connection for forwarding
This removes creation of the gRPC connection for forwarding on a
per-forward basis, and initializes a single connection when a Server is
initialized.  This seems to follow the intended usage pattern of gRPC
more consistently.

I was a little torn on weather or not to add a separate function to
initialize (lazily) the gRPC connection.  I decided to just use a
property instead, as it makes the flushing implementation cleaner and I
think it better communicates that the connection is specifically
initialized at server creation time. I decided to put this into
(*Server).Start, as it seems a little more of a natural place than the
constructor (for network initialization).