Pre-initialize gRPC connection for proxying
The proxy was previously initializing a new gRPC connection each time it
went to perform a forward to a downstream host.  This change modifies
that behavior so that it pre-initializes a connection beforehand to each
host.

To implement this, I added a thread-safe map container type that handles
storing and creating gRPC connections per host.  I tried to keep the
locking pretty straightforward, but it definitely adds a lot of
complexity to the proxy.  Existing tests and some new tests that I added
verify that it is working correctly.