proxysrv: fix data races with error handling
The changes in 29b862e missed a pretty obvious race condition related to
error handling in the proxies.  Specifically, it was possible for
both concurrent append operations as well as concurrent trace client
writes, both of which are not thread-safe.  This reworks some of the
error handling so that this race is no longer possible, by sending
errors over a channel rather than directly appending to the array.

I also fixed some race conditions in the test, which were also
triggering the race detector.