Correctly handle flushing the broken socket
Turns out it was possible that a flush could be kicked off but could
not reach the flushBackend, returning yet another ErrWouldBlock (which
counted as an error, but broke assumptions further down). Now, we loop
until both the outer and inner flush have succeeded, which should
ensure this test succeeds on Travis.