Fix an ineffectual assignment in proxysrv
It was correctly pointed out that (*proxysrv.Server).Serve tries to
handle errors when closing it's TCP listener, but in reality errors
returned by closing the listener would always have overriden those
returned normally from the function (specifically from the gRPC server).
This is due to the fact that the error was being directly returned, not
set to the local "err" variable.

This change just modified the gRPC Serve function to set the "err"
variable directly, then return it.  This should fix the previously
described condition.