proxysrv: Preallocate static tag maps
proxysrv.Server contained a couple of areas where it created a static
map of tags (to submit with SSF metrics) for every request.  This is a
pretty pointless allocation, so I moved these to be preallocated as a
package level variable.  There's nothing particularly specific about the
maps to the proxysrv.Server type, so it seemed a little more natural to
me to just set these at the package level.