Add a separate parameter for gRPC forwarding
This change adds a separate parameter "grpc_forward_address" which will
configure gRPC forwarding as opposed to HTTP.  The "forward_use_grpc"
option that I previously introduced is removed in favor of the two
separate parameters. "forward_address" and "grpc_forward_address" are
exclusive options, so the server will throw an error if they are both
set.

I retained the set of struct fields that I previously added to Server,
specifically the "forwardUseGRPC" parameter, as well as using
"ForwardAddr" for both addresses.  I personally think this will be a
little easier in the long run if HTTP forwarding is ever removed, as
there will be less changes to make.  Using a simple boolean switch also
seems a little more straighforward to use to me, rather than checking
which string parameter is set.