importsrv: batch forwarded metrics to workers
This converts importsrv to send all metrics destined for a worker in a
single batch, rather than each one individually.  This purpose is to
reduce the total number of channel sends and receives, which appear to
have a much larger cost than I originally thought.  This article talks
about an amazingly similar scenario where Datadog found this to be a
problem while processing metrics:
http://jmoiron.net/blog/go-performance-tales

I also converted the "Worker" to ingest metrics in batches rather than
one-by-one.

Unfortunately this didn't seem to have a huge impact on performance.
The benchmark for the "SendMetrics" route of importsrv is pretty
synthetic, as it does send the metrics over a channel but nothing
actually reads it.  The benchmarks didn't show any huge speedups
unfortunately.