Unflake SSF end-to-end tests
This test submits two custom metric packets with a purpose:testing tag.
The server is expected to build an array of metrics and flush it out on
a channel. We confirm that the array has both of the expected metrics in
it.

However, sometimes the flush will happen between the first and second
metric packets. The flushed array only contains one packet so the test
fails. The server then shuts down before the second metric can be sent,
which results in an "Ignoring Accept error while shutting down" error
message. This message is proof of the flake's root cause, which is fixed
in this commit.