proxysrv: periodically report active handler count
We should be tracking the number of goroutines currently handling proxy
RPCs, as the number of these are currently unbounded.  This change adds
an atomic counter that contains the count of goroutines handling proxy
RPCs, as well as new functionality to proxysrv.Server so it will
periodically report this count as a gauge.  The metric produced will be
"veneur_proxy.proxy.active_goroutines", and this should make it possible
to at least see when a proxy has spawned too many goroutines to handle
RPCs.

Additionally, I also added a default forwarding timeout of 10 seconds to
the "sendMetrics" handler.  This default is not present on the HTTP
version, but I think it makes sense to enforce it in the gRPC server
since we have the ability to set it from the beginning.