importsrv: track the different parts of the handler
This adds new metrics that will cause "import.response_duration_ns" to
be tagged with different regions of the gRPC handler.  This mirrors the
functionality of the HTTP implementation, which also does this.

The goal of this change is to be able to get better visibility into
performance issues with the gRPC handler.