Add a metricpb package
The metricpb package contains protobuf definitions and generated code to
serialize metrics to protobufs.  The top-level type exported is the
Metric, which is effectively equivalent to a JSONMetric.  This type
contains the "key" of the metric (name, type, and tags), as well as a
value.  The value is implemented as a protobuf "oneof" field, which
allows the different values of different metric types to all be used.

I then added a "value" type for each metric type.  Both CounterValue and
GaugeValue ended up being the same, but I decided to keep them separate
to make their wire definition more flexible in the future.  This way it
will be easy to selectively add fields that are forwarded to each type
as necessary.

I also added a "testutils" package, with some super simple helpers for
generating random metrics.  These are used later in tests and
benchmarks.