Add a gRPC proxy package and related protobufs (#439)
* Add protobuf support to tdigest

This adds support for converting a MergingDigest to a
protobuf-compatible struct, MergingDigestData.  This type will be used
serializing a Histogram to a protobuf.

I added some helper methods to make it easier to work with this new
type:
* NewMergingFromData: Initializes a MergingDigest from a
MergingDigestData.  Used when deserializing a protobuf-encoded metric.
* (*MergingDigest).Data: Called when serializing a metric to a protobuf.

* Add a metricpb package

The metricpb package contains protobuf definitions and generated code to
serialize metrics to protobufs.  The top-level type exported is the
Metric, which is effectively equivalent to a JSONMetric.  This type
contains the "key" of the metric (name, type, and tags), as well as a
value.  The value is implemented as a protobuf "oneof" field, which
allows the different values of different metric types to all be used.

I then added a "value" type for each metric type.  Both CounterValue and
GaugeValue ended up being the same, but I decided to keep them separate
to make their wire definition more flexible in the future.  This way it
will be easy to selectively add fields that are forwarded to each type
as necessary.

I also added a "testutils" package, with some super simple helpers for
generating random metrics.  These are used later in tests and
benchmarks.

* Add forwardrpc (forward gRPC service definition)

The forwardrpc package defines the common gRPC service to send metrics
from one Veneur to another.  It uses the metricpb.(*Metric) type to
actually serialize the metrics, and defines a single RPC to send many of
them in a batch.

I considered implementing this as a client-side streaming RPC, but that
honestly seems like a bit of a premature optimization.  From what I
could tell it would seriously complicate the code for a pretty
unmeasurable performance gain (at least right now).  It may be that this
should be replaced in the future if profiling shows that it would be
fast.

I also added a forwardtest package, which is effectively httptest but
specifically for the Forward service.  This is later used in a couple
different places to greatly simplify test cases around handling of these
RPCs.

* Add a proxy gRPC server (proxysrv)

The proxysrv package includes a gRPC server that implements the
forwardrpc.Forward service.  This basically implements the same logic as
the current proxy server, but in a separate package (and obviously with
gRPC).  It uses a consistent hash ring for forwarding metrics, and then
forwards to each one over gRPC.

I added tests as appropriate using the forwardtest mock gRPC server.

* Update Gopkg.lock to include new deps

* Fix incorrect use of atomic counter and add tests

I didn't realize that atomic.LoadInt64 should be used to read the value
of an atomic counter.  This change uses that call as opposed to just
reading the raw value, and also adds some tests for the counters.

* Stop handling defer'ed errors in proxysrv

Trying to handle errors from calls in "defer" for (proxysrv).Server was
really just causing more problems than it was worth.  I removed this
functionality, which simplifies the code a little bit.

* Add better comments for (proxysrv).forwardError