importsrv: reduce allocations from hashing metrics
From profiling global Veneurs in a production environment, we see that
the garbage collector is generally dominating the CPU usage.  Digging in
a little more, it looks like the process to hash a metric to a
destination worker in the "importsrv" package is actually doing a ton of
allocations.

To optimize this, I pulled in an external fnv hash implementation that
handles strings with zero allocations.  I also removed usages of
"MetricKey", as it does a lot of string manipulations that perform a ton
of allocations, in favor of just writing strings directly into the hash.

The benchmarks show nice improvements in the number of allocations, see
https://gist.github.com/noahgoldman/034fb2a52aadd60c383321fe9d3a4ba3 for
a comparison of benchmark results. There is a 2x reduction for a small
number (10 metrics) and a 6x reduction for 100 metrics, which is closer
to our average of around 180 currently.