Proof of concept: send internal metrics directly without conversion
Currently, veneur's internal metrics reporting happens like this:
Unless they're wrapped in a span, they get wrapped in a span, then
that span is sent down the span channel; then a span worker picks up
that span and sends it to all span sinks, and eventually the
ssfmetrics span sink gets the span, picks out the metrics samples and
reports them.

This could be much easier / happen with much allocation and
server-side work: Instead of wrapping and sending a span, just send
the batch of metrics and process that by injecting the SSFSamples as
translated into UDPMetrics into the metrics workers directly.

My guess/intense hope is that this will take a lot of pressure off the
span processing pipeline, relieving it from having to try and process
a span (expensive) in order to process only a few metrics contained in
the span (less expensive).