Test multiple shutdowns of the HTTP server
The new functionality of the (*Server).Serve function has the potential
to call both the HTTP and gRPC servers shutdown methods after that
server has already been stopped. The gRPC server already has a test that
verifies that the shutdown method can be called after the server is
stopped, but the HTTP server did not.

I added such a test in this change.  In order to facilitate writing this
test, I added a new "atomic boolean" to Server which will track the
number of active HTTP servers.  This allows us to call (*Server).Serve
in a separate goroutine in tests, and then later determine when the HTTP
server is actually listening (which is a little bit after the goroutine
is forked).  This avoid having to hardcode a sleep into the tests (which
would either be slow or have the potential to cause flakes).