Use the gRPC import server (#454)
* Use the gRPC import server

This utilizes the changes in https://github.com/stripe/veneur/pull/442
to allow Veneur to listen for forwarded metrics over gRPC.  The gRPC
server is started at the same time as the HTTP server (if both are
enabled), and works in a very similar way.  Metrics are received by the
"importsrv" package, then processed by workers in the new
"ImportMetricGRPC" method that I added.

Veneurs should be able to selectively enable both the gRPC and HTTP
servers individually.  The gRPC server should be disabled by default
unless the "grpc_address" configuration parameter is set.

To go along with this change, I added a bunch of new methods to the
"samplers" package to allow the different types to be converted to and
from their protobuf-equivalents. They are basically equivalent to the
current methods that deal with "JSONMetrics":
* `Metric`: Creates a metricpb.Metric, similar to `Export`.
* `Merge`: Combines the value with that of another sampler, like
`Combine`.

* Fix 'block'->'blocks' in server

* Test multiple shutdowns of the HTTP server

The new functionality of the (*Server).Serve function has the potential
to call both the HTTP and gRPC servers shutdown methods after that
server has already been stopped. The gRPC server already has a test that
verifies that the shutdown method can be called after the server is
stopped, but the HTTP server did not.

I added such a test in this change.  In order to facilitate writing this
test, I added a new "atomic boolean" to Server which will track the
number of active HTTP servers.  This allows us to call (*Server).Serve
in a separate goroutine in tests, and then later determine when the HTTP
server is actually listening (which is a little bit after the goroutine
is forked).  This avoid having to hardcode a sleep into the tests (which
would either be slow or have the potential to cause flakes).