Fix #454 incorrectly editing the v5 changelog (#492)
Back in the day when I created #454, v5 was still unreleased and I had
edited the appropriate section in the CHANGELOG.  By the time we merged
it yesterday, v5 had already been tagged and the changes had no longer
made it into that release.

This updates the changelog to show that gRPC import support was only
added for v6.  I also added in @quantcast to the changelog, as a lot of
the work done for these features was on their time.