Do not stream stdout/err in veneur-emit, pass fd's through directly
The way we did streaming output in veneur-emit previously was slightly
flawed: We seem to have introduced a race condition on our CI boxes
that caused veneur-emit to occasionally exit because io.Copy returned
with an EAGAIN. We're stumped why that would happen.

So, instead of streaming the output manually, let's do the properly
UNIXy thing and pass all FDs through to the child process as-is - now
including Stdin, too.