Forward metrics over gRPC (#493)
* Forward metrics over gRPC

This adds support for forwarding metrics over gRPC from one Veneur to
another.  A new method is added to Server (forwardGRPC) that performs
the bulk of the work along with some new helper methods on the
WorkerMetric type.

I added a new configuration option that allows gRPC forwarding to be
activated, specifically `forward_use_grpc`.  This is then used in the
Server initialization to setup a gRPC connection to the forwarding
destination.

With this change, forwarding over gRPC is now possible from local
Veneurs to a single global Veneur!  The full high-availability approach
of local Veneurs with Proxies and Globals is still not possible, as
additional changes are needed to the Proxies.

* Fix small formatting and spelling errors

This fixes an instance of 'Weather' -> 'Whether', adds a period to the
comment on "appendExportedMetric" and changes the method definition to
be all on a single line.